// src/lib/application/files/react-admin/src/components/TinyEditor/tinymce.d.ts
interface DomQueryConstructor {
  readonly prototype: DomQuery
  attrHooks: Record<string, Hook>
  cssHooks: Record<string, Hook>
  fn: DomQuery
  find: any
  expr: {
    cacheLength: number
    createPseudo: Function
    match: Record<string, RegExp>
    attrHandle: {}
    find: Record<string, Function>
    relative: Record<
      string,
      {
        dir: string
        first?: boolean
      }
    >
    preFilter: Record<string, Function>
    filter: Record<string, Function>
    pseudos: Record<string, Function>
  }
  extend: Tools['extend']
  isArray: Tools['isArray']
  new <T extends Node = Node>(
    selector?: DomQueryInitSelector<T>,
    context?: Node
  ): DomQuery<T>
  <T extends Node = Node>(
    selector?: DomQueryInitSelector<T>,
    context?: Node
  ): DomQuery<T>
  overrideDefaults(
    callback: () => {
      context: Node
      element: Element
    }
  ): DomQueryConstructor
  makeArray<T>(object: T): T[]
  inArray<T>(item: {}, array: T[]): number
  each<T>(obj: T[], callback: (i: number, value: T) => void): void
  each<T>(obj: T, callback: (key: string, obj: T[keyof T]) => void): void
  trim(str: string): string
  grep<T>(array: T[], callback: (item: any, i: number) => boolean): T[]
  unique<T>(results: T[]): T[]
  text(elem: Node): string
  contains(context: any, elem: Node): boolean
  filter(expr: string, elems: Node[], not?: boolean): any
}
// ...
interface NotificationManager {
  open: (spec: NotificationSpec) => NotificationApi
  close: () => void
  getNotifications: () => NotificationApi[]
}
// ...
interface EditorUi extends EditorUiApi {
  registry: Registry
  styleSheetLoader: StyleSheetLoader
}
// ...
interface EditorSettings extends BaseEditorSettings {
  external_plugins: Record<string, string>
  plugins: string
}
interface ParamTypeMap {
  'hash': Record<string, string>
  'string': string
  'number': number
  'boolean': boolean
  'string[]': string[]
  'array': any[]
}
interface BlobInfoImagePair {
  image: HTMLImageElement
  blobInfo: BlobInfo
}
declare class NodeChange {
  private readonly editor
  private lastPath
  constructor(editor: Editor)
  nodeChanged(args?: any): void
  private isSameElementPath
}
interface SelectionOverrides {
  showCaret: (
    direction: number,
    node: Element,
    before: boolean,
    scrollIntoView?: boolean
  ) => Range | null
  showBlockCaretContainer: (blockCaretContainer: Element) => void
  hideFakeCaret: () => void
  destroy: () => void
}
interface Quirks {
  refreshContentEditable(): void
  isHidden(): boolean
}
// ...
interface Annotator {
  register: (name: string, settings: AnnotatorSettings) => void
  annotate: (name: string, data: DecoratorData) => void
  annotationChanged: (name: string, f: AnnotationListenerApi) => void
  remove: (name: string) => void
  getAll: (name: string) => Record<string, Element[]>
}
// ...
interface DomParser {
  schema: Schema
  addAttributeFilter: (
    name: string,
    callback: (nodes: AstNode[], name: string, args: ParserArgs) => void
  ) => void
  getAttributeFilters: () => ParserFilter[]
  addNodeFilter: (
    name: string,
    callback: (nodes: AstNode[], name: string, args: ParserArgs) => void
  ) => void
  getNodeFilters: () => ParserFilter[]
  filterNode: (node: AstNode) => AstNode
  parse: (html: string, args?: ParserArgs) => AstNode
}
// ...
interface DomSerializer extends DomSerializerImpl {}
// ...
declare class EditorCommands {
  private readonly editor
  private selectionBookmark
  private commands
  constructor(editor: Editor)
  execCommand(command: string, ui?: boolean, value?: any, args?: any): boolean
  queryCommandState(command: string): boolean
  queryCommandValue(command: string): string
  addCommands<K extends keyof Commands>(commandList: Commands[K], type: K): void
  addCommands(commandList: Record<string, EditorCommandsCallback>): void
  addCommand(
    command: string,
    callback: EditorCommandCallback,
    scope?: any
  ): void
  queryCommandSupported(command: string): boolean
  addQueryStateHandler(
    command: string,
    callback: () => boolean,
    scope?: any
  ): void
  addQueryValueHandler(
    command: string,
    callback: () => string,
    scope?: any
  ): void
  hasCustomCommand(command: string): boolean
  private execNativeCommand
  private isFormatMatch
  private toggleFormat
  private storeSelection
  private restoreSelection
  private setupCommands
}
// ...
declare type Untranslated = Primitive | TokenisedString | RawString
declare type TranslatedString = string
// ...
declare class URI {
  static parseDataUri(uri: string): {
    type: string
    data: string
  }
  static isDomSafe(
    uri: string,
    context?: string,
    options?: SafeUriOptions
  ): boolean
  static getDocumentBaseUrl(loc: {
    protocol: string
    host?: string
    href?: string
    pathname?: string
  }): string
  source: string
  protocol: string
  authority: string
  userInfo: string
  user: string
  password: string
  host: string
  port: string
  relative: string
  path: string
  directory: string
  file: string
  query: string
  anchor: string
  settings: URISettings
  constructor(url: string, settings?: URISettings)
  setPath(path: string): void
  toRelative(uri: string): string
  toAbsolute(uri: string, noHost?: boolean): string
  isSameOrigin(uri: URI): boolean
  toRelPath(base: string, path: string): string
  toAbsPath(base: string, path: string): string
  getURI(noProtoHost?: boolean): string
}
interface EditorManager extends Observable<EditorManagerEventMap> {
  $: DomQueryConstructor
  defaultSettings: RawEditorSettings
  majorVersion: string
  minorVersion: string
  releaseDate: string
  editors: Editor[]
  activeEditor: Editor
  focusedEditor: Editor
  settings: RawEditorSettings
  baseURI: URI
  baseURL: string
  documentBaseURL: string
  i18n: I18n
  suffix: string
  add(this: EditorManager, editor: Editor): Editor
  addI18n: (code: string, item: Record<string, string>) => void
  createEditor(
    this: EditorManager,
    id: string,
    settings: RawEditorSettings
  ): Editor
  execCommand(
    this: EditorManager,
    cmd: string,
    ui: boolean,
    value: any
  ): boolean
  get(this: EditorManager): Editor[]
  get(this: EditorManager, id: number | string): Editor
  init(this: EditorManager, settings: RawEditorSettings): Promise<Editor[]>
  overrideDefaults(
    this: EditorManager,
    defaultSettings: Partial<RawEditorSettings>
  ): void
  remove(this: EditorManager): void
  remove(this: EditorManager, selector: string | Editor): Editor | void
  setActive(this: EditorManager, editor: Editor): void
  setup(this: EditorManager): void
  translate: (text: Untranslated) => TranslatedString
  triggerSave: () => void
  _setBaseUrl(this: EditorManager, baseUrl: string): void
}
interface EditorObservable extends Observable<EditorEventMap> {
  bindPendingEventDelegates(this: Editor): void
  toggleNativeEvent(this: Editor, name: string, state: boolean): any
  unbindAllNativeEvents(this: Editor): void
}
interface UploadResult$1 {
  element: HTMLImageElement
  status: boolean
  blobInfo: BlobInfo
  uploadUri: string
}
declare type UploadCallback = (results: UploadResult$1[]) => void
interface EditorUpload {
  blobCache: BlobCache
  addFilter: (filter: (img: HTMLImageElement) => boolean) => void
  uploadImages: (callback?: UploadCallback) => Promise<UploadResult$1[]>
  uploadImagesAuto: (
    callback?: UploadCallback
  ) => void | Promise<UploadResult$1[]>
  scanForImages: () => Promise<BlobInfoImagePair[]>
  destroy: () => void
}
// ...
interface EditorMode {
  isReadOnly: () => boolean
  set: (mode: string) => void
  get: () => string
  register: (mode: string, api: EditorModeApi) => void
}
interface EditorModeApi {
  activate: () => void
  deactivate: () => void
  editorReadOnly: boolean
}
interface Plugin {
  getMetadata?: () => {
    name: string
    url: string
  }
  [key: string]: any
}
declare type PluginManager = AddOnManager<Plugin>
interface ShortcutsConstructor {
  readonly prototype: Shortcuts
  new (editor: Editor): Shortcuts
}
declare type CommandFunc = string | [string, boolean, any] | (() => void)
declare class Shortcuts {
  private readonly editor
  private readonly shortcuts
  private pendingPatterns
  constructor(editor: Editor)
  add(pattern: string, desc: string, cmdFunc: CommandFunc, scope?: any): boolean
  remove(pattern: string): boolean
  private normalizeCommandFunc
  private createShortcut
  private hasModifier
  private isFunctionKey
  private matchShortcut
  private executeShortcutAction
}
// ...
declare class Editor implements EditorObservable {
  documentBaseUrl: string
  baseUri: URI
  settings: EditorSettings
  id: string
  plugins: Record<string, Plugin>
  documentBaseURI: URI
  baseURI: URI
  contentCSS: string[]
  contentStyles: string[]
  ui: EditorUi
  mode: EditorMode
  setMode: (mode: string) => void
  $: DomQueryConstructor
  shortcuts: Shortcuts
  loadedCSS: Record<string, any>
  editorCommands: EditorCommands
  suffix: string
  editorManager: EditorManager
  inline: boolean
  isNotDirty: boolean
  callbackLookup: any
  _nodeChangeDispatcher: NodeChange
  editorUpload: EditorUpload
  annotator: Annotator
  bodyElement: HTMLElement
  bookmark: any
  composing: boolean
  container: HTMLElement
  contentAreaContainer: HTMLElement
  contentDocument: Document
  contentWindow: Window
  delegates: Record<string, (event: any) => void>
  destroyed: boolean
  dom: DOMUtils
  editorContainer: HTMLElement
  eventRoot?: Element
  formatter: Formatter
  formElement: HTMLElement
  formEventDelegate: (e: Event) => void
  hasHiddenInput: boolean
  hasVisual: boolean
  hidden: boolean
  iframeElement: HTMLIFrameElement | null
  iframeHTML: string
  initialized: boolean
  notificationManager: NotificationManager
  orgDisplay: string
  orgVisibility: string
  parser: DomParser
  quirks: Quirks
  readonly: boolean
  removed: boolean
  schema: Schema
  selection: EditorSelection
  serializer: DomSerializer
  startContent: string
  targetElm: HTMLElement
  theme: Theme
  undoManager: UndoManager
  validate: boolean
  windowManager: WindowManager
  _beforeUnload: () => void
  _eventDispatcher: EventDispatcher<NativeEventMap>
  _mceOldSubmit: any
  _pendingNativeEvents: string[]
  _selectionOverrides: SelectionOverrides
  _skinLoaded: boolean
  bindPendingEventDelegates: EditorObservable['bindPendingEventDelegates']
  toggleNativeEvent: EditorObservable['toggleNativeEvent']
  unbindAllNativeEvents: EditorObservable['unbindAllNativeEvents']
  fire: EditorObservable['fire']
  on: EditorObservable['on']
  off: EditorObservable['off']
  once: EditorObservable['once']
  hasEventListeners: EditorObservable['hasEventListeners']
  constructor(
    id: string,
    settings: RawEditorSettings,
    editorManager: EditorManager
  )
  render(): void
  focus(skipFocus?: boolean): void
  hasFocus(): boolean
  execCallback(name: string, ...x: any[]): any
  translate(text: Untranslated): TranslatedString
  getParam<K extends keyof ParamTypeMap>(
    name: string,
    defaultVal: ParamTypeMap[K],
    type: K
  ): ParamTypeMap[K]
  getParam<K extends keyof EditorSettings>(
    name: K,
    defaultVal?: EditorSettings[K],
    type?: string
  ): EditorSettings[K]
  getParam<T>(name: string, defaultVal: T, type?: string): T
  hasPlugin(name: string, loaded?: boolean): boolean
  nodeChanged(args?: any): void
  addCommand(
    name: string,
    callback: EditorCommandCallback,
    scope?: object
  ): void
  addQueryStateHandler(name: string, callback: () => boolean, scope?: any): void
  addQueryValueHandler(name: string, callback: () => string, scope?: any): void
  addShortcut(
    pattern: string,
    desc: string,
    cmdFunc: string | [string, boolean, any] | (() => void),
    scope?: any
  ): void
  execCommand(cmd: string, ui?: boolean, value?: any, args?: any): boolean
  queryCommandState(cmd: string): boolean
  queryCommandValue(cmd: string): string
  queryCommandSupported(cmd: string): boolean
  show(): void
  hide(): void
  isHidden(): boolean
  setProgressState(state: boolean, time?: number): void
  load(args?: any): string
  save(args?: any): string
  setContent(content: string, args?: SetContentArgs): string
  setContent(content: AstNode, args?: SetContentArgs): AstNode
  setContent(content: Content, args?: SetContentArgs): Content
  getContent(
    args: {
      format: 'tree'
    } & GetContentArgs
  ): AstNode
  getContent(args?: GetContentArgs): string
  insertContent(content: string, args?: any): void
  resetContent(initialContent?: string): void
  isDirty(): boolean
  setDirty(state: boolean): void
  getContainer(): HTMLElement
  getContentAreaContainer(): HTMLElement
  getElement(): HTMLElement
  getWin(): Window
  getDoc(): Document
  getBody(): HTMLElement
  convertURL(url: string, name: string, elm?: any): string
  addVisual(elm?: HTMLElement): void
  remove(): void
  destroy(automatic?: boolean): void
  uploadImages(callback?: UploadCallback): Promise<UploadResult$1[]>
  _scanForImages(): Promise<BlobInfoImagePair[]>
  addButton(): void
  addSidebar(): void
  addMenuItem(): void
  addContextToolbar(): void
}